+++ tests/unit/test_batch_study.py
@@ -1,6 +1,7 @@
 """
 Tests for the batch_study.py
 """
+import os
 import pybamm
 import unittest
 
@@ -46,10 +47,7 @@ class TestBatchStudy(unittest.TestCase):
         # Tests for exceptions
         for name in pybamm.BatchStudy.INPUT_LIST:
             with self.assertRaises(ValueError):
-                pybamm.BatchStudy(
-                    models={"SPM": spm, "DFN": dfn},
-                    **{name: {None}}
-                )
+                pybamm.BatchStudy(models={"SPM": spm, "DFN": dfn}, **{name: {None}})
 
         # Tests for None when only models are given with permutations=False
         bs_false_only_models.solve(t_eval=[0, 3600])
@@ -68,9 +66,7 @@ class TestBatchStudy(unittest.TestCase):
             self.assertIn(output_model, models_list)
 
             output_solver = bs_false.sims[num].solver.name
-            solvers_list = [
-                solver.name for solver in bs_false.solvers.values()
-            ]
+            solvers_list = [solver.name for solver in bs_false.solvers.values()]
             self.assertIn(output_solver, solvers_list)
 
             output_experiment = bs_false.sims[
@@ -108,8 +104,10 @@ class TestBatchStudy(unittest.TestCase):
         bs_false.solve()
         bs_false.plot(testing=True)
         self.assertEqual(2, len(bs_false.sims))
+
+        sols = [sim.solution for sim in bs_false.sims]
         for num in range(len(bs_false.sims)):
-            output_model = bs_false.sims[num].all_models[0].name
+            output_model = sols[num].all_models[0].name
             models_list = [model.name for model in bs_false.models.values()]
             self.assertIn(output_model, models_list)
 
@@ -117,10 +115,25 @@ class TestBatchStudy(unittest.TestCase):
         bs_true.solve()
         bs_true.plot(testing=True)
         self.assertEqual(4, len(bs_true.sims))
+
+        sols = [sim.solution for sim in bs_true.sims]
         for num in range(len(bs_true.sims)):
-            output_model = bs_true.sims[num].all_models[0].name
+            output_model = sols[num].all_models[0].name
             models_list = [model.name for model in bs_true.models.values()]
             self.assertIn(output_model, models_list)
+
+    def test_create_gif(self):
+        bs = pybamm.BatchStudy({"spm": pybamm.lithium_ion.SPM()})
+        bs.solve([0, 3600])
+
+        # create a GIF before calling the plot method
+        bs.create_gif(number_of_images=5, duration=1)
+
+        # create a GIF after calling the plot method
+        bs.plot(testing=True)
+        bs.create_gif(number_of_images=5, duration=1)
+
+        os.remove("plot.gif")
 
 
 if __name__ == "__main__":